requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
tqdm>=4.66.0
openai>=1.3.0
neo4j>=5.13.0
fastapi>=0.104.0
uvicorn>=0.24.0
jinja2>=3.1.0
python-multipart>=0.0.6
python-dotenv>=1.1.0 